requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
chardet==2.3.0
cryptography==1.7.1
Flask-Login==0.4.1
flask-mongoengine==0.9.5
Flask-RESTful==0.3.6
Flask-WTF==0.14.2
httplib2==0.9.2
idna==2.2
itsdangerous==0.24
Jinja2==2.10
kombu==4.1.0
ldap3==2.5
MarkupSafe==1.0
redis==2.10.6
requests==2.12.4
SQLAlchemy==1.2.6
urllib3==1.22
Werkzeug==0.14.1
WTForms==2.1
Flask-JWT
